backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.0
motor==3.6.0
pydantic==2.8.2
python-dotenv==1.0.1
